# Conflicts:
#	restcomm/restcomm.application/src/main/webapp/WEB-INF/data/hsql/restcomm.script
#	restcomm/restcomm.application/src/main/webapp/WEB-INF/scripts/mariadb/sql/clients.xml
#	restcomm/restcomm.application/src/main/webapp/WEB-INF/sql/clients.xml
#	restcomm/restcomm.dao/src/main/java/org/mobicents/servlet/restcomm/dao/ClientsDao.java
#	restcomm/restcomm.dao/src/main/java/org/mobicents/servlet/restcomm/dao/mybatis/MybatisClientsDao.java
#	restcomm/restcomm.dao/src/test/resources/data/restcomm.script
#	restcomm/restcomm.dao/src/test/resources/restcomm.script
#	restcomm/restcomm.http/src/main/java/org/mobicents/servlet/restcomm/http/converter/AbstractConverter.java
#	restcomm/restcomm.sms/src/main/java/org/mobicents/servlet/restcomm/sms/SmsService.java
#	restcomm/restcomm.telephony/src/main/java/org/mobicents/servlet/restcomm/telephony/ua/UserAgentManager.java
#	restcomm/restcomm.testsuite/src/test/resources/org/mobicents/servlet/restcomm/ussd/restcomm.script_ussdPullTest
#	restcomm/restcomm.testsuite/src/test/resources/restcomm.script
#	restcomm/restcomm.testsuite/src/test/resources/restcomm.script_SmsTest
#	restcomm/restcomm.testsuite/src/test/resources/restcomm.script_accounts_test
#	restcomm/restcomm.testsuite/src/test/resources/restcomm.script_dialActionTest
#	restcomm/restcomm.testsuite/src/test/resources/restcomm.script_dialTest
#	restcomm/restcomm.testsuite/src/test/resources/restcomm.script_playTest
#	restcomm/restcomm.testsuite/src/test/resources/restcomm_with_Data.script

+++ restcomm/restcomm.commons/src/main/java/org/mobicents/servlet/restcomm/amazonS3/S3AccessTool.java
@@ -28,6 +28,9 @@ import java.util.Date;
 
 import javax.activation.MimetypesFileTypeMap;
 
+import com.amazonaws.regions.Region;
+import com.amazonaws.regions.Regions;
+import org.apache.commons.io.FileUtils;
 import org.apache.log4j.Logger;
 
 import com.amazonaws.AmazonClientException;
@@ -54,12 +57,13 @@ public class S3AccessTool {
     private String securityKey;
     private String bucketName;
     private String folder;
+    private String bucketRegion;
     private boolean reducedRedundancy;
     private int daysToRetainPublicUrl;
     private boolean removeOriginalFile;
 
     public S3AccessTool(final String accessKey, final String securityKey, final String bucketName, final String folder,
-            final boolean reducedRedundancy, final int daysToRetainPublicUrl, final boolean removeOriginalFile) {
+            final boolean reducedRedundancy, final int daysToRetainPublicUrl, final boolean removeOriginalFile,final String bucketRegion) {
         this.accessKey = accessKey;
         this.securityKey = securityKey;
         this.bucketName = bucketName;
@@ -67,11 +71,14 @@ public class S3AccessTool {
         this.reducedRedundancy = reducedRedundancy;
         this.daysToRetainPublicUrl = daysToRetainPublicUrl;
         this.removeOriginalFile = removeOriginalFile;
+        this.bucketRegion = bucketRegion;
     }
 
     public URI uploadFile(final String fileToUpload) {
-        AWSCredentials credentials = new BasicAWSCredentials(accessKey, securityKey);
+        AWSCredentials credentials =new BasicAWSCredentials(accessKey, securityKey);
         AmazonS3 s3client = new AmazonS3Client(credentials);
+        s3client.setRegion(Region.getRegion(Regions.fromName(bucketRegion)));
+        logger.info("S3 Region: "+bucketRegion.toString());
         try {
             StringBuffer bucket = new StringBuffer();
             bucket.append(bucketName);
@@ -101,19 +108,25 @@ public class S3AccessTool {
             URL downloadUrl = s3client.generatePresignedUrl(generatePresignedUrlRequestGET);
 
             //Second upload the file to S3
-            while (!file.exists()){}
-            PutObjectRequest putRequest = new PutObjectRequest(bucket.toString(), file.getName(), file);
-            ObjectMetadata metadata = new ObjectMetadata();
-            metadata.setContentType(new MimetypesFileTypeMap().getContentType(file));
-            putRequest.setMetadata(metadata);
-            if (reducedRedundancy)
-                putRequest.setStorageClass(StorageClass.ReducedRedundancy);
-            s3client.putObject(putRequest);
+//            while (!file.exists()){}
+            while (!FileUtils.waitFor(file, 30)){}
+            if (file.exists()) {
+                PutObjectRequest putRequest = new PutObjectRequest(bucket.toString(), file.getName(), file);
+                ObjectMetadata metadata = new ObjectMetadata();
+                metadata.setContentType(new MimetypesFileTypeMap().getContentType(file));
+                putRequest.setMetadata(metadata);
+                if (reducedRedundancy)
+                    putRequest.setStorageClass(StorageClass.ReducedRedundancy);
+                s3client.putObject(putRequest);
 
-            if(removeOriginalFile) {
-                removeLocalFile(file);
+                if (removeOriginalFile) {
+                    removeLocalFile(file);
+                }
+                return downloadUrl.toURI();
+            } else {
+                logger.error("Timeout waiting for the recording file: "+file.getAbsolutePath());
+                return null;
             }
-            return downloadUrl.toURI();
          } catch (AmazonServiceException ase) {
             logger.error("Caught an AmazonServiceException");
             logger.error("Error Message:    " + ase.getMessage());
